Convert string input into unsigned int and compare
We need to convert the string into an u32 int so that we can compare it
against the `secret_number` that we generate randomly.

We shadow the `guess` variable when converting.

    let guess: u32 = guess.trim().parse()

The second `guess` in this line refers to the value that is input whilst
the first `guess` is what will "overwritten".